Removed config.py
